% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_VR_object.R
\name{add_VR_object}
\alias{add_VR_object}
\title{Add a 3D Object to the VR Environment}
\usage{
add_VR_object(
  vr_env,
  obj_path,
  position = c(0, 0, 0),
  rotation = c(0, 0, 0),
  scale = c(0.01, 0.01, 0.01),
  adjust_to_table = TRUE
)
}
\arguments{
\item{vr_env}{A list representing the VR environment, containing the
\code{html_content} as a character string that holds the HTML for the VR scene.}

\item{obj_path}{A character string specifying the path to the OBJ file to be added.}

\item{position}{A numeric vector of length 3 specifying the position of the
object in the VR scene, given as \verb{(x, y, z)}. Defaults to \verb{(0, 0, 0)}.}

\item{rotation}{A numeric vector of length 3 specifying the rotation of the
object in the VR scene, given as \verb{(x, y, z)} angles in degrees. Defaults to \verb{(0, 0, 0)}.}

\item{scale}{A numeric vector of length 3 specifying the scale of the object
in the VR scene. Defaults to \verb{(0.01, 0.01, 0.01)}.}

\item{adjust_to_table}{A logical indicating whether to automatically include
table under the object. Defaults to \code{TRUE}.}
}
\value{
A list representing the updated VR environment, with the newly added
object and modified \code{html_content}.
}
\description{
This function adds a 3D object specified by an OBJ file to an existing
VR environment. The object is positioned, rotated, and scaled according
to the provided parameters. Optionally, it can also adjust the position
of a table in relation to the object.
}
\details{
The function first checks if the specified OBJ file exists and constructs
an MTL file path based on the OBJ file's name. It creates appropriate
\verb{<a-asset-item>} tags for the object and material and appends these to
the VR environment's HTML content before the closing \verb{</a-scene>} tag.
If the table adjustment feature is enabled, the relevant JavaScript component
is added to the HTML as well.
}
\examples{
# Example of adding a 3D object to a VR environment
vr_env <- list(html_content = "<a-scene></a-scene>")
obj_path <- "path/to/your/object.obj"
updated_vr_env <- add_VR_object(vr_env, obj_path, position = c(1, 0, -3), 
                                 rotation = c(0, 180, 0), scale = c(0.1, 0.1, 0.1))

}
