% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_aoi.R
\name{define_aoi}
\alias{define_aoi}
\title{Define Area of Interest}
\usage{
define_aoi(lon, lat, buffer_distance)
}
\arguments{
\item{lon}{Longitude of the central point.}

\item{lat}{Latitude of the central point.}

\item{buffer_distance}{Distance for the buffer around the central point in meters.}
}
\value{
A bounding box string for the area of interest.
}
\description{
Define Area of Interest
}
