% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_3Dbag_items.R
\name{get_3Dbag_items}
\alias{get_3Dbag_items}
\title{Get 3DBag Items Data}
\usage{
get_3Dbag_items(bbox)
}
\arguments{
\item{bbox}{A bounding box string defining the area of interest.}
}
\value{
A list containing the parsed JSON data from 3DBag.
}
\description{
Get 3DBag Data
}
