% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_3Dbag_buildings.R
\name{plot_3Dbag_buildings}
\alias{plot_3Dbag_buildings}
\title{Plot 3DBag Buildings}
\usage{
plot_3Dbag_buildings(lon, lat, buffer_distance)
}
\arguments{
\item{lon}{Longitude of the central point.}

\item{lat}{Latitude of the central point.}

\item{buffer_distance}{Distance for the buffer around the central point in meters.}
}
\value{
A matrix of final coordinates used for 3D plotting.
}
\description{
Plot 3DBag Buildings
}
