% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_VR.R
\name{rotate_VR}
\alias{rotate_VR}
\title{Add Rotation to VR Model}
\usage{
rotate_VR(input_html, rotate, speed = 1e-04, clockwise = 1)
}
\arguments{
\item{input_html}{Path to the input HTML file.}

\item{rotate}{Logical value to enable/disable rotation. If TRUE, rotation is enabled.}

\item{speed}{(Optional) Numeric value for rotation speed. Defaults to 0.0001.}

\item{clockwise}{(Optional) Numeric value to set rotation direction. 1 for clockwise, 0 for counterclockwise. Defaults to 1.}
}
\value{
Updates the input HTML file with the specified rotation settings for the 3D model.
}
\description{
This function updates an existing VR HTML file to add or remove a rotation component for the 3D model based on the provided parameters.
}
\examples{
\dontrun{
rotate_VR("input.html", TRUE)
rotate_VR("input.html", TRUE, speed = 0.0002, clockwise = 0)
rotate_VR("input.html", FALSE)
}
}
