% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_model_as_glb.R
\name{save_model_as_glb}
\alias{save_model_as_glb}
\title{Save 3D Model as GLB File}
\usage{
save_model_as_glb(filepath, filename, coords, color = "blue")
}
\arguments{
\item{filepath}{The path where the GLB file will be saved.}

\item{filename}{The name of the GLB file to save (without extension).}

\item{coords}{A data frame with columns x, y, and z.}

\item{color}{The color to be used for the 3D model. Default is "blue".}
}
\description{
Save 3D Model as GLB File
}
