% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_camera_position.R
\name{set_camera_position}
\alias{set_camera_position}
\title{Set Camera Position in a VR Environment}
\usage{
set_camera_position(vr_env, camera_position = c(0, 2, 5))
}
\arguments{
\item{vr_env}{A list containing the HTML content of the VR environment. The list should have an element named \code{html_content}.}

\item{camera_position}{A numeric vector of length 3 specifying the camera's position in the format \code{c(x, y, z)}. Defaults to \code{c(0, 2, 5)}.}
}
\value{
A modified list containing the updated HTML content of the VR environment.
}
\description{
This function sets or updates the position of the camera in a given VR environment.
}
