% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{validColor}
\alias{validColor}
\title{Validate Color}
\usage{
validColor(x)
}
\arguments{
\item{color}{A string representing the color (hex code or named color).}
}
\value{
TRUE if the color is valid, otherwise FALSE.
}
\description{
This function checks whether a given color is valid. It supports hex codes and a set of basic named colors.
}
\examples{
validColor("#FF5733")  # Returns TRUE
validColor("red")      # Returns TRUE
validColor("invalidColorName")  # Returns FALSE

}
