% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_VR_environment.R
\name{set_VR_environment}
\alias{set_VR_environment}
\title{Set VR Environment}
\usage{
set_VR_environment(
  input_html,
  environment_preset,
  skyType = "atmosphere",
  skyColor = "blue"
)
}
\arguments{
\item{input_html}{Path to the input HTML file.}

\item{environment_preset}{The environment preset to use for the A-Frame environment component.}

\item{skyType}{(Optional) The type of sky. Valid values: "color", "gradient", "atmosphere". Defaults to "atmosphere".}

\item{skyColor}{(Optional) The sky color. When skyType is "color" or "gradient", it sets the main sky color. Defaults to "blue".}
}
\value{
Updates the input HTML file with the specified environment preset and sky settings for the 3D model.
}
\description{
Set VR Environment
}
\details{
This function updates an existing VR HTML file to add an A-Frame environment component with specified parameters.
}
\examples{
\dontrun{
set_VR_environment("output.html", "forest")
set_VR_environment("output.html", "egypt", skyType = "color", skyColor = "red")
set_VR_environment("output.html", "japan", skyType = "gradient", skyColor = "#FFDD99")
}
}
